API consistency batch: usize counts, honest NanoTime conversions, dead RunFor::done
All five items of #834, taken now because three are semver-major and
9.0.0 is unpublished -- after bump.yml they would cost a 10.0.0.

**RunFor::done deleted.** Zero callers anywhere, and it was wrong:
`cycle > *cycles` against Kernel::begin_cycle's `cycles >= end`. Aligning
it and routing the kernel through it would have been worse -- the kernel
does not evaluate RunFor per cycle at all, it resolves the variant once in
Kernel::build into end_time/end_cycle sentinels and compares integers
after that, so a predicate would mean re-matching the enum every cycle on
the hot path to restore a copy nobody uses. RunFor now documents
begin_cycle as its single evaluator.

**Filter's unread condition tick flag removed.** In<'a> is now
((&'a T, bool), &'a bool). Verified in the derive rather than assumed:
parse_in_shape sets edge_uses_tick[1] = false for a bare &'a bool, so
expand_builder emits a constant instead of a per-activation
__ticked.borrow()[up]. Active/passive classification comes from
args.passive, not the flag, so the condition edge still activates the
node -- which is where the documented resample behaviour actually comes
from. Sole such site in the tree; every other op reads every flag it
declares.

**Duration -> NanoTime hoisted to start** for Throttle, Delay,
DelayWithReset and Window (which converted in both places).

**limit widened u32 -> usize**, with skip alongside it -- see below.

**NanoTime conversions made honest.** From<i64>, From<f64> and From<u128>
were bare `as` casts the compiler could insert at any .into(), so
NanoTime::from(-1i64) silently became MAX -- an instant 584 years in the
future. All three are deleted with **no replacement**.

Named _lossy constructors were written first and then removed: nothing in
the tree called the i64 or f64 ones, and each had to invent a policy
(negatives to the epoch, NaN to the epoch) on behalf of a caller who is
better placed to choose. NanoTime is a u64 count since the epoch, so a
caller holding a signed or floating count should decide what out-of-range
means for their feed and go through NanoTime::new with a checked
conversion, which is total. The u128 narrowing survives as a private
from_nanos_u128_saturating because Duration::as_nanos forces it -- a
Duration is 64 bits of seconds plus 32 of nanos, wider than this type --
making it an implementation detail of From<Duration>, not vocabulary.

From<Duration> is KEPT, contrary to the issue: it is not truncating.
Duration and NanoTime are both nanosecond counts and the conversion is
exact across NanoTime's whole range; its only edge was overflow past ~584
years, which now saturates instead of wrapping.

Sub is documented, not changed. A saturating Sub returns 0, which is
indistinguishable from a genuine zero-elapsed latency measurement -- it
would convert a loud debug panic into a plausible wrong number.

Two things this surfaced that the issue did not predict:

- Overriding start obliges every op generic to appear in Cfg or State.
  #[op] emits a real start forwarder only when start is overridden, and
  the compiled tiers call it with nothing but Cfg/State to infer from, so
  adding start to Throttle broke every nitro! call site with E0282.
  ThrottleState<T> carries a PhantomData<T> to anchor it.
- DelayWithResetFwd needed start delegated too. Without it every
  delay_with_reset silently degrades to a pass-through, and an
  interpreted-vs-compiled parity assertion cannot catch that, because all
  three tiers reach the op through the same forwarder and break
  identically. Confirmed by deleting the delegation and watching the new
  test fail.

skip was widened alongside limit though #834 does not mention it: it sits
beside limit in fluent.rs with the identical u32 Cfg/State, so fixing one
and leaving the other reproduces exactly the inconsistency item 4 exists
to remove. Three lines, no Python binding, literal callers only.

docs/release-notes/9.0.0.md claimed RunMode, RunFor and NanoTime were
unchanged, which these removals falsify -- corrected, with all three
removals and the limit/skip widening recorded under "What is gone".

Three rules folded back into the /new-op skill per CLAUDE.md's
living-document rule: the tick-flag cost, start-obliges-generic-anchoring,
and the delegating-forwarder trap with its parity-assertions-cannot-catch-
this caveat.

+++ crates/wingfoil-python/src/graph.rs
@@ -423,7 +423,7 @@ impl PyStream {
     }
 
     /// Pass through the first `limit` values, then stay quiet.
-    pub fn limit(&self, limit: u32) -> PyStream {
+    pub fn limit(&self, limit: usize) -> PyStream {
         self.wrap(self.stream.limit(limit))
     }
